shaii types refactor
