+++ src/types/index.ts
@@ -1,0 +1,6 @@
+export * from "./anime.types";
+export * from "./commands.types";
+export * from "./database.types";
+export * from "./imageProcessing.types";
+export * from "./textProcessing.types";
+export * from "./waifu.types";
