+++ src/types/textProcessing.types.ts
@@ -1,0 +1,4 @@
+export type TextProcessorFn = (sentence: string, ...args: any) => Promise<string>;
+export interface TextProcessors {
+  [key: string]: TextProcessorFn;
+}
